moved internal packages code and testdata to internal structure
