+++ internal/processing/refs_test.go
@@ -1,7 +1,8 @@
-package main
+package processing
 
 import (
 	"bytes"
+	"log"
 	"os"
 	"testing"
 )
@@ -11,7 +12,12 @@ var (
 )
 
 func readFile(relativePath string) []byte {
-	data, _ := os.ReadFile(TEST_RESOURCES_PATH + relativePath)
+	path := TEST_RESOURCES_PATH + relativePath
+	data, err := os.ReadFile(path)
+	if err != nil {
+		log.Printf("Error during reading a file from path '%s'", path)
+		panic(err)
+	}
 	return data
 }
 
@@ -21,7 +27,7 @@ func TestGatherHRefs(t *testing.T) {
 		path := "/Better_Motherfucking_Website.html"
 		data := readFile(path)
 		reader := bytes.NewReader(data)
-		hrefs, _ := gatherHRefs(reader)
+		hrefs, _ := GatherHRefs(reader)
 		if len(hrefs) != 6 {
 			t.Errorf("From '%s' returned '%d' number of refs, wanted '%d'", path, len(hrefs), 6)
 		}
